requirements.txt
pandas
numpy
requests
scikit-learn
matplotlib
seaborn
plotly
streamlit
python-dotenv
scipy
